agregando nueva libreria de serializacion y ordenando un poco el codigo

+++ app.py
@@ -1,7 +1,7 @@
 import os
 from flask import Flask, jsonify, request
 from flask_sqlalchemy import SQLAlchemy
-from marshmallow import Schema, fields
+from json_serializable import JSONSerializable
 from books import Books
 
 dbdir = 'sqlite:///' + os.path.abspath(os.getcwd()) + '/db_books.db'
@@ -10,6 +10,9 @@ app = Flask(__name__)
 app.config['SQLALCHEMY_DATABASE_URI'] = dbdir
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 db = SQLAlchemy(app)
+JSONSerializable(app)
+
+
 
 class Books(db.Model):
 	id = db.Column(db.Integer, primary_key=True)
@@ -19,12 +22,6 @@ class Books(db.Model):
 	launching = db.Column(db.String(100))
 
 
-class BookSchema(Schema):
-	id = fields.Int()
-	title = fields.Str()
-	description = fields.Str()
-	author = fields.Str()
-	launching = fields.Str()
 
 @app.route('/')
 def index():
@@ -35,25 +32,28 @@ def index():
 @app.route('/api/books', methods=["GET"])
 def  get_books():
 	books_list = Books.query.all()
-	
-	book_schema = BookSchema(many=True)
-	results = book_schema.dump(books_list)
 
-	return jsonify(results)
+	return jsonify(books_list)
 
 
 #Buscar libro
 @app.route('/api/books/<int:book_id>', methods=["GET"])
 def search_book(book_id):
-	book_found = Books.query.filter_by(id=book_id).first()
+	
+	if request.method == 'GET':
+		book_found = Books.query.filter_by(id=book_id).first()
 
-	print("query done", book_found.title)
-	return jsonify(
-		{"title":book_found.title,
-		"description":book_found.description,
-		"author":book_found.author,
-		"launching":book_found.launching}
-	)
+		return jsonify(
+			{
+				"title":book_found.title,
+				"description":book_found.description,
+				"author":book_found.author,
+				"launching":book_found.launching
+			}
+		)
+	
+	else:
+		return jsonify({"message": "Book not found"})
 
 
 
@@ -63,10 +63,10 @@ def add_book():
 
 	if request.method == 'POST':
 		new_book = Books(
-			title=request.json["title"],
-			description=request.json["description"],
-			author=request.json["author"],
-			launching=request.json["launching"]
+			title				= request.json["title"],
+			description = request.json["description"],
+			author			= request.json["author"],
+			launching		= request.json["launching"]
 		)
 	
 		db.session.add(new_book)
@@ -81,35 +81,45 @@ def add_book():
 #Eliminar un libro
 @app.route('/api/books/<int:book_id>', methods=["DELETE"])
 def remove_book(book_id):
-	book_found = Books.query.filter_by(id=book_id).first()
 
-	db.session.delete(book_found)
-	db.session.commit()
+	if request.method == 'POST':
+		book_found = Books.query.filter_by(id=book_id).first()
 
-	print("query done", book_found.title)
-	return jsonify({"message":"Book deleted"})
+		db.session.delete(book_found)
+		db.session.commit()
+
+		return jsonify({"message":"Book deleted"})
+	
+	else:
+		return jsonify({"message": "Book not found"})
 
 
 #Actulizar o editar un libro
 @app.route('/api/books/<int:book_id>', methods=["GET", "PUT"])
 def update_book(book_id):
-	book_found = Books.query.filter_by(id=book_id).first()
-		
-	book_found.title      = request.json["title"]
-	book_found.description= request.json["description"]
-	book_found.author     = request.json["author"]
-	book_found.launching  = request.json["launching"]
 
-	db.session.add(book_found)
-	db.session.commit()
+	if request.method == 'PUT':
+		book_found = Books.query.filter_by(id=book_id).first()
+	
+		book_found.title       = request.json["title"]
+		book_found.description = request.json["description"]
+		book_found.author      = request.json["author"]
+		book_found.launching   = request.json["launching"]
 
-	print("Book updated", book_found.title)
-	return jsonify(
-		{"title": book_found.title,
-     "description": book_found.description,
-     "author": book_found.author,
-     "launching": book_found.launching}
-	)
+		db.session.add(book_found)
+		db.session.commit()
+
+		return jsonify(
+			{
+				"title": book_found.title,
+				"description": book_found.description,
+				"author": book_found.author,
+				"launching": book_found.launching
+			}
+		)
+	
+	else:
+		return jsonify({"message": "Book not found"})
 
 
 if __name__ == "__main__":
